Add fake host support (assuming your ircd supports it)
Add definitions in various places to support assigning fake hosts to
users.  We will not give you (or make for you) patches to make your
ircd support this feature, so please do not ask.  You must set a
certain srvx.conf setting to be able to assign fakehosts; finding it
is left as an exercise to the reader.
git-archimport-id: srvx@srvx.net--2004-srvx/srvx--devo--1.3--patch-67