bahamut: Add support for NICKIPSTR
This patch adds support for recognizing and utilizing NICKIPSTR
which allows Bahamut to process IP addresses as strings. As such,
calling irc_pton with an empty irc_in_addr_t caused the stringified
IP address to be :: which Bahamut rejected, forcing us to add local
users with 127.0.0.1 (which we statically initialize).

Finally, irc_user checked for the first byte of the users's nick to be
a space which caused my local copy to fail to add local users properly,
so I swapped that logic out with IsDummy which seems to have resolved
the issue.